Remove all ice_servers on native sync WebRTC cameras
